HSEARCH-2073 Move serialization testing helper to the Engine module
