+++ engine/src/test/java/org/hibernate/search/testsupport/serialization/SerializationTestHelper.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.test;
+package org.hibernate.search.testsupport.serialization;
 
 import static org.junit.Assert.assertEquals;
 
@@ -16,7 +16,7 @@ import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.hibernate.search.test.SerializationTestHelper.Foo.TestInnerClass;
+import org.hibernate.search.testsupport.serialization.SerializationTestHelper.Foo.TestInnerClass;
 import org.junit.Test;
 
 /**
