+++ orm/src/test/java/org/hibernate/search/test/jpa/EntityManagerSerializationTest.java
@@ -13,9 +13,9 @@ import org.junit.Test;
 
 import org.hibernate.search.jpa.FullTextEntityManager;
 import org.hibernate.search.jpa.Search;
-import org.hibernate.search.test.SerializationTestHelper;
 import org.hibernate.search.testsupport.TestConstants;
 import org.hibernate.search.testsupport.TestForIssue;
+import org.hibernate.search.testsupport.serialization.SerializationTestHelper;
 
 import static org.junit.Assert.assertEquals;
